Khôi - Admin dashboard, review user, ...

+++ admin/delete_product.php
@@ -2,10 +2,13 @@
 require_once("../config/config.php");
 require_once("auth_admin.php");
 
-$id = $_POST['id'];
+if (!isset($_GET['id'])) {
+    die("Thiếu ID");
+}
 
-mysqli_query($conn,"DELETE FROM products WHERE id=$id");
+$id = (int) $_GET['id'];
 
-echo json_encode([
-"status"=>"success"
-]);
+mysqli_query($conn, "DELETE FROM products WHERE id = $id");
+
+header("Location: pages/manage_products.php");
+exit;
